+++ lib/core/index.ts
@@ -9,3 +9,4 @@ export * from './id';
 export * from './iterator';
 export * from './result';
 export * from './value-object';
+export * from './create-many-domain-instance';