+++ CHANGE_LOG.md
@@ -5,6 +5,39 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 ---
+### 1.10.0 - 2022-08-07
+
+### Added
+
+- createMany method on domain entity and value objects
+
+```ts
+
+
+const { result, data } = ValueObject.createMany([
+  Class<AgeProps>(Age, props),
+  Class<PriceProps>(Price, props),
+  Class<NameProps>(Name, props)
+]);
+
+result.isOk() // true
+
+const age = data.next() as IResult<Age>;
+const price = data.next() as IResult<Price>;
+const name = data.next() as IResult<Name>;
+
+age.value().get('value') // 21
+
+```
+
+### Change
+
+### Result:
+
+- from `isOK()` to `isOk()`
+- from `OK()` to `Ok()`
+
+---
 
 ### 1.9.0-beta - 2022-08-06
 
@@ -14,7 +47,7 @@ All notable changes to this project will be documented in this file.
 
 ```ts
 
-const result: Result<void> = Result.OK();
+const result: Result<void> = Result.Ok();
 
 ```
 - Rename methods: